Reorganized file structure of tests
