+++ package.json
@@ -33,8 +33,8 @@
 		"lint": "./node_modules/.bin/eslint --config=config/.eslintrc.json --ext ts src",
 		"watch": "./node_modules/.bin/tsc -watch -p ./config/",
 		"pretest": "npm run clean && npm run compile && npm run lint",
-		"unit-test": "node ./out/unit-tests/runTest.js",
-		"e2e-test": "./node_modules/.bin/extest setup-and-run -u ./out/e2e-tests/*.test.js",
+		"unit-test": "node ./out/tests/unit-tests/runTest.js",
+		"e2e-test": "./node_modules/.bin/extest setup-and-run -u ./out/tests/e2e-tests/*.test.js",
 		"docs": "./node_modules/.bin/jsdoc -c ./config/jsdoc.json"
 	},
 	"devDependencies": {
