+++ tests/unit-tests/runTest.ts
@@ -6,7 +6,7 @@ async function main() {
     try {
         // The folder containing the Extension Manifest package.json
         // Passed to `--extensionDevelopmentPath`
-        const extensionDevelopmentPath = path.resolve(__dirname, "../../src/");
+        const extensionDevelopmentPath = path.resolve(__dirname, "../../../src/");
 
         // The path to test runner
         // Passed to --extensionTestsPath
